Authorize using gates

+++ app/Http/Controllers/QuestionController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\AskQuestionRequest;
 use App\Question;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Http\Request;
 
 class QuestionController extends Controller
@@ -65,6 +66,10 @@ class QuestionController extends Controller
      */
     public function edit(Question $question)
     {
+        //using gate as authorization
+        if(Gate::denies('update-question', $question)) {
+            abort(403, "Access Denied! You are not allowed to do this operation!");
+        }
         return view('questions.edit', compact('question'));
     }
 
@@ -77,6 +82,10 @@ class QuestionController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if(Gate::denies('update-question', $question)) {
+            abort(403, "Access Denied! You are not allowed to do this operation!");
+        }
+
         $question->update($request->only('title', 'body'));
 
         return redirect()->route('questions.index')->with('success', "Your question has been updated!");
@@ -90,6 +99,10 @@ class QuestionController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(Gate::denies('delete-question', $question)) {
+            abort(403, "Access Denied! You are not allowed to do this operation!");
+        }
+
         $question->delete();
         return redirect('/questions')->with('success', "Question has been deleted.");
     }
